Added error handling to scripts

+++ deploy.sh
@@ -2,43 +2,52 @@
 
 ./zola-linux-0.15.3 build
 
-# Create JS file from search data
-mv ./public/search-data/index.html ./public/js/search-data.js
-rm -rf ./public/search-data/
+if [ $? -eq 0 ]
+then
+  echo "Successfully created file"
+  exit 0
 
-# Remove taxonomy pages
-rm -rf ./public/tags ./public/category
+  # Create JS file from search data
+  mv ./public/search-data/index.html ./public/js/search-data.js
+  rm -rf ./public/search-data/
 
-# Remove taxonomy from the sitemap
-awk 'NR==FNR{if (/muffinman.io\/(tags|category)\/.*/) for (i=-1;i<=1;i++) del[NR+i]; next} !(FNR in del)' ./public/sitemap.xml ./public/sitemap.xml > ./public/sitemap-tmp.xml
+  # Remove taxonomy pages
+  rm -rf ./public/tags ./public/category
 
-# Remove search data from the sitemap
-awk 'NR==FNR{if (/muffinman.io\/search-data/) for (i=-1;i<=1;i++) del[NR+i]; next} !(FNR in del)' ./public/sitemap-tmp.xml ./public/sitemap-tmp.xml > ./public/sitemap.xml
+  # Remove taxonomy from the sitemap
+  awk 'NR==FNR{if (/muffinman.io\/(tags|category)\/.*/) for (i=-1;i<=1;i++) del[NR+i]; next} !(FNR in del)' ./public/sitemap.xml ./public/sitemap.xml > ./public/sitemap-tmp.xml
 
-# Remove temp
-rm ./public/sitemap-tmp.xml
+  # Remove search data from the sitemap
+  awk 'NR==FNR{if (/muffinman.io\/search-data/) for (i=-1;i<=1;i++) del[NR+i]; next} !(FNR in del)' ./public/sitemap-tmp.xml ./public/sitemap-tmp.xml > ./public/sitemap.xml
 
-# Switch to gh-pages branch
-git fetch origin gh-pages gh-pages
-git checkout gh-pages
+  # Remove temp
+  rm ./public/sitemap-tmp.xml
 
-git config user.name "GitHub Actions"
-git config user.email "github-actions-bot@users.noreply.${GITHUB_HOSTNAME}"
+  # Switch to gh-pages branch
+  git fetch origin gh-pages gh-pages
+  git checkout gh-pages
 
-# Delete old build
-find . ! -path './.git' ! -path . ! -name 'public' -maxdepth 1 -exec rm -rf {} +
+  git config user.name "GitHub Actions"
+  git config user.email "github-actions-bot@users.noreply.${GITHUB_HOSTNAME}"
 
-# Copy everything from the new build to root
-cp -r ./public/* ./
-# Remove public folder
-rm -rf ./public
+  # Delete old build
+  find . ! -path './.git' ! -path . ! -name 'public' -maxdepth 1 -exec rm -rf {} +
 
-# Create .nojekyll file
-touch .nojekyll
+  # Copy everything from the new build to root
+  cp -r ./public/* ./
+  # Remove public folder
+  rm -rf ./public
 
-# Commit with current time
-git add .
-git commit -a -m "Deploy `date +'%Y-%m-%d %H:%M:%S'`"
+  # Create .nojekyll file
+  touch .nojekyll
 
-# Push changes
-git push origin gh-pages
+  # Commit with current time
+  git add .
+  git commit -a -m "Deploy `date +'%Y-%m-%d %H:%M:%S'`"
+
+  # Push changes
+  git push origin gh-pages
+else
+  echo "Zola build failed"
+  exit 1
+fi
